begun list selection functionality

+++ lists.cpp
@@ -1,8 +1,13 @@
 #include "lists.h"
 
+#include <cstdint>
 #include <string>
 #include <fstream>
 #include <filesystem>
+#include <iostream>
+#include <sstream>
+#include "ansi.h"
+#include "win.h"
 
 /**
  * Initialises file structure for loading, and saving lists.
@@ -33,6 +38,8 @@ bool Lists::init()
  */
 bool Lists::save(const char* filename, const std::vector<std::string>& list)
 {
+	if (std::filesystem::exists(filename)) return false;
+
 	std::ofstream outFile = std::ofstream(filename, std::ios::out | std::ios::binary);
 	if (!outFile) return false;
 
@@ -54,6 +61,8 @@ bool Lists::save(const char* filename, const std::vector<std::string>& list)
  */
 bool Lists::load(const char* filename, std::vector<std::string>& list)
 {
+	if (!std::filesystem::exists(filename)) return false;
+
 	std::ifstream inFile = std::ifstream(filename, std::ios::in | std::ios::binary);
 	if (!inFile) return false;
 
@@ -70,5 +79,53 @@ bool Lists::load(const char* filename, std::vector<std::string>& list)
 	}
 
 	inFile.close();
+	return true;
+}
+
+/**
+ * Lists all files in the lists directory.
+ * @return true if successful, false otherwise
+ */
+bool Lists::list()
+{
+	std::filesystem::path listsPath("./Lists");
+
+	std::ostringstream formattedStream;
+	short index = 0;
+
+	std::vector<std::string> listOfLists;
+
+	for (const std::filesystem::directory_entry& list : std::filesystem::directory_iterator(listsPath)) {
+		if (!list.is_regular_file()) continue;
+		if (list.path().extension() != ".bin") continue;
+
+		formattedStream << Fore::YELLOW << "[" << Style::UNDER << std::to_string(index) << Style::RESET <<
+			Fore::YELLOW << "]" << Style::RESET << " " << list.path().stem().string();
+
+		listOfLists.push_back(formattedStream.str());
+
+		formattedStream.str("");
+		formattedStream.clear();
+
+		index++;
+	}
+
+	index = 0;
+
+	for (const std::string& string : listOfLists) {
+		short padding = 8;
+
+		if (index + 3 == viewportDepth) {
+			short extras = listOfLists.size() - index;
+			std::cout << Erase::LINE << std::string(padding, ' ') << Fore::BLACK << "... (+" <<
+				std::to_string(extras) << (extras == 1 ? " extra)" : " extras)") << Style::RESET << std::endl;
+
+			break;
+		}
+
+		std::cout << Erase::LINE << std::string(padding, ' ') << string << std::endl;
+		index++;
+	}
+
 	return true;
 }
